+++ src/main/java/Graphic/Goompa.java
@@ -16,7 +16,7 @@ public class Goompa extends Enemy {
     private int y;
     private int width = 60;
     private int height = 65;
-    private int velocity = -5;
+    private double velocity = -5;
 
     public Goompa(int xx, int yy) {
         this.setSize(width, height);
@@ -89,11 +89,11 @@ public class Goompa extends Enemy {
         this.height = height;
     }
 
-    public int getVelocity() {
+    public double getVelocity() {
         return velocity;
     }
 
-    public void setVelocity(int velocity) {
+    public void setVelocity(double velocity) {
         this.velocity = velocity;
     }
 }
